requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-WTF
psycopg2-binary
python-dotenv
email_validator
flask-socketio
pytest
pytest-flask